+++ help50/__init__.py
@@ -16,5 +16,30 @@ except DistributionNotFound:
 else:
     __version__ = _dist.version
 
-from ._api import *
 
+import collections
+
+HELPERS = collections.defaultdict(list)
+PREPROCESSORS = collections.defaultdict(list)
+
+
+def helper(*domains):
+    """ Decorator that indicateas that a given function is a helper. """
+    def decorator(func):
+        for domain in domains:
+            HELPERS[domain].append(func)
+        return func
+    return decorator
+
+
+def preprocessor(*domains):
+    """ Decorator that indicateas that a given function is a preprocessor. """
+    def decorator(func):
+        for domain in domains:
+            PREPROCESSORS[domain].append(func)
+        return func
+    return decorator
+
+
+class Error(Exception):
+    pass
